Fixed missing pragma once in Item.h files, updated stack_v2.h interface to be consistent for all Ex4_19 implementations, all stack_v3.h implementations now return bool parameters for functions that may fail indicating success or not. Implemented Ex4_24. Added a test file to all Exercises from Ex4.19 through to Ex4.24

+++ Chapter4/PushdownStack/Exercise/Ex4_19/Ex4_19b/stackList_v2.c
@@ -48,7 +48,7 @@
      return x;
  }
  
- void STACKinit(int maxN) {
+ void STACKinit(size_t maxN) {
      //reset stack
      head = nullptr;
  }
